+++ src/test/java/org/broadinstitute/hellbender/engine/ReferenceMultiSourceUnitTest.java
@@ -4,7 +4,7 @@ import com.google.cloud.dataflow.sdk.options.PipelineOptions;
 import org.apache.spark.SparkConf;
 import org.broadinstitute.hellbender.engine.datasources.ReferenceMultiSource;
 import org.broadinstitute.hellbender.engine.datasources.ReferenceWindowFunctions;
-import org.broadinstitute.hellbender.engine.spark.SparkTestUtils;
+import org.broadinstitute.hellbender.utils.test.SparkTestUtils;
 import org.broadinstitute.hellbender.utils.test.BaseTest;
 import org.testng.Assert;
 import org.testng.annotations.Test;
