+++ src/test/java/org/broadinstitute/hellbender/engine/spark/SAMRecordSerializerUnitTest.java
@@ -6,6 +6,7 @@ import org.apache.spark.SparkConf;
 import org.apache.spark.serializer.KryoRegistrator;
 import org.broadinstitute.hellbender.utils.read.ArtificialReadUtils;
 import org.broadinstitute.hellbender.utils.read.SAMRecordToGATKReadAdapter;
+import org.broadinstitute.hellbender.utils.test.SparkTestUtils;
 import org.testng.Assert;
 import org.testng.annotations.Test;
